Clone script takes directory

+++ scripts/clone.js
@@ -1,6 +1,11 @@
 const { rmdir, mkdir } = require("fs");
 const { exec } = require("child_process");
 
+const directory = process.argv[2];
+if (!directory) {
+  throw new Error("Please provide a directory");
+}
+
 rmdir("./copy", () => {
   mkdir("./copy", () => {
     exec(
@@ -14,18 +19,18 @@ rmdir("./copy", () => {
         exec(
           "yarn",
           {
-            cwd: "./copy/flights-microfrontends/header",
+            cwd: `./copy/flights-microfrontends/${directory}`,
           },
           () => {
-            console.log("DONE INSTALLING DEPENDENCIES");
+            console.log(`DONE INSTALLING DEPENDENCIES TO ${directory}`);
 
             exec(
               "yarn deploy",
               {
-                cwd: "./copy/flights-microfrontends/header",
+                cwd: `./copy/flights-microfrontends/${directory}`,
               },
               () => {
-                console.log("DONE DEPLOYING");
+                console.log(`DONE DEPLOYING TO ${directory}`);
               }
             );
           }
